Add WiFi disconnection feature
This commit adds a new method to the WiFiHandler class that allows the device to disconnect from the current WiFi network. The new method, disconnect(), checks if the device is currently connected and disconnects if it is. It also prints a message indicating whether the disconnection was successful or if the device was not connected.

End of commit message